==> 08-Bibliografia.Rmd <==
\backmatter

`r if (knitr::is_html_output()) '
# Bibliografía {-}
'`

```{r include=FALSE}
knitr::write_bib(c(
  .packages(), 'bookdown', 'knitr', 'rmarkdown', 'htmlwidgets', 'webshot', 'DT',
  'miniUI', 'tufte', 'servr', 'citr', 'rticles'
), 'packages.bib')
```
